+++ include/ponder/classbuilder.hpp
@@ -40,12 +40,13 @@
 #include <ponder/detail/constructorimpl.hpp>
 #include <ponder/detail/propertyfactory.hpp>
 #include <ponder/pondertype.hpp>
+#include <ponder/userdata.hpp>
 #include <cassert>
 #include <string>
 
 
 namespace ponder {
-    
+        
 /**
  * \brief Proxy class which fills a metaclass with its members
  *
@@ -101,8 +102,7 @@ public:
      * \code
      * struct Point
      * {
-     *     float x;
-     *     float y;
+     *     float x, y;
      *
      *     float length() const;
      * };
@@ -140,7 +140,6 @@ public:
      * class Entity
      * {
      * public:
-     *
      *     Point p;
      * };
      *
@@ -182,7 +181,7 @@ public:
      */
     template <typename... A>
     ClassBuilder<T>& constructor();
-
+    
     /**
      * \brief Add properties and/or functions from an external source
      *
@@ -201,7 +200,6 @@ public:
      * class MyClassMapper
      * {
      * public:
-     *
      *     MyClassMapper();
      *
      *     size_t propertyCount();
@@ -225,29 +223,32 @@ public:
     template <template <typename> class U>
     ClassBuilder<T>& external();
 
+    /**
+     * \brief Add user data to the last declared member type
+     *
+     * \code
+     * ponder::Class::declare<MyClass>("MyClass")
+     *     .function("foo", &MyClass::foo)( ponder::UserData("user", 3) );
+     * \endcode
+     *
+     * \return Reference to this, in order to chain other calls
+     */
+    template <typename... U>
+    ClassBuilder<T>& operator () (U&&... uds)
+    {
+        const std::initializer_list<UserData> il = {uds...};
+        for (UserData const& ud : il)
+            userDataStore()->setValue(*m_currentType, ud.getName(), ud.getValue());
+        return *this;
+    }
+
 private:
 
-    /**
-     * \brief Add a new property to the target class
-     *
-     * \param property Property to add
-     *
-     * \return Reference to this, in order to chain other calls
-     */
     ClassBuilder<T>& addProperty(Property* property);
-
-    /**
-     * \brief Add a new function to the target class
-     *
-     * \param function Function to add
-     *
-     * \return Reference to this, in order to chain other calls
-     */
     ClassBuilder<T>& addFunction(Function* function);
 
-    Class* m_target; ///< Target metaclass to fill
-    Property* m_currentProperty; ///< Last metaproperty which has been declared
-    Function* m_currentFunction; ///< Last function which has been declared
+    Class* m_target; // Target metaclass to fill
+    Type* m_currentType; // Last member type which has been declared
 };
 
 } // namespace ponder