+++ include/ponder/enum.hpp
@@ -76,17 +76,18 @@ namespace ponder
  *
  * \sa Class, EnumBuilder
  */
-class PONDER_API Enum : detail::noncopyable
+class PONDER_API Enum : public Type
 {
+    PONDER__NON_COPYABLE(Enum);
 public:
-    typedef long EnumValue;         //!< Type used to hold the enum value.
+    typedef long EnumValue;  //!< Type used to hold the enum value.
 
     /**
      * \brief Structure defining the <name, value> pairs stored in metaenums
      */
     struct Pair {
-        Id name;    //!< Enum name
-        EnumValue value;            //!< Enum value
+        Id name;            //!< Enum name
+        EnumValue value;    //!< Enum value
         
         /**
          * \brief Constructor.
@@ -107,8 +108,6 @@ public:
         E valueAs() const {return static_cast<E>(value);}
     };
     
-public:
-
     /**
      * \brief Declare a new metaenum
      *