+++ test/ponder/userobject.cpp
@@ -88,8 +88,10 @@ namespace UserObjectTest
         return stream << m.x;
     }
     
-    struct MyNonCopyableClass : ponder::detail::noncopyable
+    struct MyNonCopyableClass
     {
+        MyNonCopyableClass() {}
+        PONDER__NON_COPYABLE(MyNonCopyableClass);
     };
     
     struct MyAbstractClass